Move fixable `map_unwrap_or` cases to rustfixed test
